修正了服务器测试工具的编译错误 原因是修改了项目配置

+++ source/prolib/StaticString.h
@@ -74,7 +74,7 @@ void StaticString<_STRINGLEN>::operator = ( const char* pstr)
 	if( pstr == 0)
 		this->str_length_ =0;
 	else
-		str_length_ =strlen( pstr);
+		str_length_ =(int)strlen( pstr);
 
 	if( str_length_ > _STRINGLEN)
 		str_length_ =_STRINGLEN;
